api/requirements.txt
flask
tiktoken
openai
langchain
pinecone-client